Split off object type into standalone header
