+++ src/openrct2/management/Research.h
@@ -10,8 +10,9 @@
 #pragma once
 
 #include "../common.h"
-#include "../object/Object.h"
 #include "../object/ObjectLimits.h"
+#include "../object/ObjectType.h"
+#include "../ride/RideTypes.h"
 #include "../util/Util.h"
 
 #include <optional>
