+++ src/openrct2/object/FootpathItemEntry.h
@@ -11,8 +11,7 @@
 
 #include "../common.h"
 #include "../interface/Cursors.h"
-
-using ObjectEntryIndex = uint16_t;
+#include "ObjectType.h"
 
 enum class PathBitDrawType : uint8_t
 {
